fix: modified the file structure
