+++ src/pages/profile/index.tsx
@@ -1,11 +1,10 @@
 import { useRouter } from 'next/router'
-import Image from 'next/image'
 
 import { Button, HeadMeta, getMainLayout, Typography, Loader } from '@/components'
 import { PATH } from '@/consts/route-paths'
 import { ImageOutline } from '@/assets/icons'
-import { useMeQuery } from '@/services/auth/auth.api'
-import { useGetProfileQuery } from '@/services/profile/profile.api'
+import { useMeQuery } from '@/services/auth/auth-api'
+import { useGetProfileQuery } from '@/services/profile/profile-api'
 
 import s from './profile.module.scss'
 import { useTranslation } from '@/hooks'
