+++ src/services/auth/index.ts
@@ -1,2 +1,2 @@
-export * from './auth.api'
+export * from './auth-api'
 export * from './types'
