+++ src/services/base-api.ts
@@ -1,9 +1,9 @@
-import { baseQueryWithReauth } from '@/services'
 import { createApi } from '@reduxjs/toolkit/query/react'
+import { baseQueryWithReauth } from '@/services/base-api-with-reauth'
 
 export const baseApi = createApi({
   reducerPath: 'baseApi',
-  baseQuery: baseQueryWithReauth,
   tagTypes: ['getProfile'],
+  baseQuery: baseQueryWithReauth,
   endpoints: () => ({}),
 })
