+++ src/services/index.ts
@@ -1,3 +1,2 @@
 export * from './store'
-export * from './base.api'
-export * from './base-api-with-reauth'
+export * from './base-api'
