+++ src/services/profile/index.ts
@@ -1,2 +1,2 @@
-export * from './profile.api'
+export * from './profile-api'
 export * from './types'
